requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
aiosqlite==0.19.0
python-dotenv==1.0.0
pydantic==1.10.7
alpha_vantage==2.3.1
pytest==7.3.1
pytest-asyncio==0.21.0
